+++ R/condor-package.R
@@ -23,6 +23,11 @@
 #'   \code{\link{condor_dir}}      \tab list directories\cr
 #'   \code{\link{condor_download}} \tab download
 #' }
+#' \emph{Stop and remove:}
+#' \tabular{ll}{
+#'   \code{\link{condor_rm}}    \tab stop jobs\cr
+#'   \code{\link{condor_rmdir}} \tab remove directories
+#' }
 #' \emph{Utilities:}
 #' \tabular{ll}{
 #'   \code{\link{condor_log}}         \tab show log file\cr
@@ -30,7 +35,9 @@
 #'   \code{\link{ssh_exec_stdout}}    \tab execute command
 #' }
 #'
-#' @author Arni Magnusson with contributions by Jemery Day and Thomas Teears.
+#' @author
+#' Arni Magnusson and Nan Yao, with contributions by Jemery Day and Thomas
+#' Teears.
 #'
 #' @references
 #' \url{https://github.com/PacificCommunity/ofp-sam-condor}